[robert] Flask: Hello world
I followed the tutorials about flask and flask-socketio and it seems
like it will suit our needs pretty nicely.

This hello world app is just a chat with no username for now but there
are ways of making sure a message is unique thanks to the session

We just need to initialiaze the session on the first request and we'll
have a unique ID for a player that the client cannot change. We might
want to have a database system to make sure a game doesn't crash and
vanish but I would like to implement the game before digging further
into socketio